Remove unused nodeKey field from get account(s) classes.
Refactor AccountWrap and AccountsWrap to share AccountContents class instead of duplicating it.
